% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRLScpp.R
\name{bw.locpolysmooth}
\alias{bw.locpolysmooth}
\title{K-fold cross validation to find optimal bandwidth for local polynomial kernel smoother}
\usage{
bw.locpolysmooth(
  Lt,
  Ly,
  method = "HUBER",
  kernel = "epanechnikov",
  delta = NULL,
  bw_cand = NULL,
  cv_loss = "HUBER",
  K = 5,
  ncores = 1,
  ...
)
}
\arguments{
\item{Lt}{a list of vectors containing time points for each curve}

\item{Ly}{a list of vectors containing observations for each curve}

\item{method}{"Huber" or "WRM" or "Bisquare"}

\item{kernel}{a kernel function for kernel smoothing ("epan", "gauss" are supported.)}

\item{delta}{cut-off value for "huber"(Huber) or "bisquare"(Tukey's biweight function).
Default is 1.345 for "huber" and 4.685 for "bisquare" for 95\% ARE.}

\item{bw_cand}{user defined bandwidth candidates for CV}

\item{cv_loss}{"Huber" or "L1" or "L2"}

\item{K}{the number of folds}

\item{ncores}{If ncores > 1, it implements \code{foreach()} in \code{doParallel} for CV.}

\item{...}{parameters are same with \code{local_kern_smooth()}.}
}
\value{
bandwidth
}
\description{
K-fold cross validation to find optimal bandwidth for local polynomial kernel smoother
}
