% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen.R
\name{check_eigen_sign}
\alias{check_eigen_sign}
\title{Change the sign of eigenvectors to target eigenvectors}
\usage{
check_eigen_sign(eig_vec, target)
}
\arguments{
\item{eig_vec}{a vector (n x 1) or matrix (n x q) containing estimated eigenvectors}

\item{target}{a vector (n x 1) or matrix (n x q) containing true (or want to convert) eigenvectors}
}
\description{
Change the sign of eigenvectors to target eigenvectors
}
