% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_Mest.R
\name{cov_Mest}
\alias{cov_Mest}
\title{Marginal M-estimator for Covaraince function (covariance estimator)}
\usage{
cov_Mest(x, smooth = FALSE, bw = NULL, make.pos.semidef = TRUE, noise.var = 0)
}
\arguments{
\item{x}{a n x p matrix or a list containing Lt and Ly}

\item{smooth}{If it is \code{TRUE}, the bivariate Nadaraya-Watson smoothing is performed for the covariance estimator. (Default is \code{FALSE}.)}

\item{bw}{an optional parameter when \code{smooth = TRUE} for a bandwidth of the bivariate smoothing.}

\item{make.pos.semidef}{a logical value whether made a covarariance estimator positive semi-definite (Default is \code{TRUE}.)}

\item{noise.var}{a numeric value of the noise variance}
}
\value{
a p x p covariance function
}
\description{
Obtain the marginal M-estimator for covariance function.
}
