% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_ogk.R
\name{cov_ogk}
\alias{cov_ogk}
\title{Robust covariance function esimation for partially observed functional data}
\usage{
cov_ogk(
  X,
  type = c("huber", "bisquare", "tdist"),
  MM = TRUE,
  smooth = TRUE,
  grid = NULL,
  bw = NULL,
  cv = FALSE,
  df = 3,
  cv_optns = list(bw_cand = NULL, K = 5, ncores = 1)
)
}
\arguments{
\item{X}{a n x p matrix. It allows NA.}

\item{type}{the option for robust dispersion estimator. "huber", "bisquare", and "tdist" are supported.}

\item{MM}{the option for M-scale estimator in GK identity. If it is FALSE, the same method using \code{type} is used, that is the iterative algorithm. The closed form solution using method of moments can be used when \code{MM == TRUE}. Defalut is TRUE.}

\item{smooth}{If it is TRUE, bivariate Nadaraya-Watson smoothing is performed using \code{fields::smooth2d()}. Default is TRUE.}

\item{grid}{a vector containing the observed timepoints}

\item{bw}{a bandwidth when \code{smooth = TRUE}.}

\item{cv}{If it is TRUE, K-fold cross-validation is performed for the bandwidth selection when \code{smooth = TRUE}.}

\item{df}{the degrees of freedm when \code{type = "tdist"}.}

\item{cv_optns}{the options of K-fold cross-validation when \code{cv = TRUE}. See Details.}
}
\value{
a list contatining as follows:
\item{mean}{the vector containing the robust mean function.}
\item{cov}{a matrix containing robust covariance function.}
\item{bw}{a bandwidth of the bivariate smoothing selected from K-fold cross-validation}
\item{cv.obj}{cv.obj from bandwidth selection}
}
\description{
The location and scale functions are computed via pointwise M-estimator, and the covariance function is obtained via robust pairwise computation based on Orthogonalized Gnanadesikan-Kettenring (OGK) estimation.
Additionally, bivariate Nadaraya-Watson smoothing is applied for smoothed covariance surfaces.
}
\details{
The options of \code{cv_optns}:
\describe{
\item{bw_cand}{a vector contains the candidates of bandwidths for bivariate smoothing.}
\item{K}{the number of folds for K-fold cross validation.}
\item{ncores}{the number of cores on \code{foreach} for parallel computing.}
}
}
\examples{
set.seed(100)
x.list <- sim_delaigle(n = 100,
                       type = "partial",
                       out.prop = 0.2,
                       out.type = 1,
                       dist = "normal")
x <- list2matrix(x.list)
cov.obj <- cov_ogk(x,
                   type = "huber",
                   bw = 0.1)
mu.ogk.sm <- cov.obj$mean
cov.ogk.sm <- cov.obj$cov

}
\references{
\cite{Park, Y., Kim, H., & Lim, Y. (2022+). Functional principal component analysis for partially observed elliptical process, Under review.}

\cite{Maronna, R. A., & Zamar, R. H. (2002). Robust estimates of location and dispersion for high-dimensional datasets. Technometrics, 44(4), 307-317.}
}
