% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov.R
\name{covfunc.rob}
\alias{covfunc.rob}
\title{Covariance estimation via semiparametric approach}
\usage{
covfunc.rob(
  Lt,
  Ly,
  newt = NULL,
  method = c("Huber", "WRM", "Bisquare"),
  mu = NULL,
  weig = NULL,
  ...
)
}
\arguments{
\item{Lt}{a list of vectors containing time points for each curve}

\item{Ly}{a list of vectors containing observations for each curve}

\item{newt}{a vector containing time points to estimate}

\item{method}{"huber", "WRM" are supported}

\item{mu}{\code{meanfunc.rob} object from \code{meanfunc.rob()}}

\item{weig}{weight}

\item{...}{additional options of \code{varfunc.rob()}}
}
\value{
a \code{covfunc.rob} object contatining as follows:
\item{sig2}{a noise variance estimate}
\item{theta}{a estiamted parameter in parametric correlation function}
\item{mu.hat}{a estimated mean function corresponded to unlist(Lt)}
\item{domain}{a range of timepoints}
\item{mu}{a \code{meanfunc.rob} object}
\item{sig2x}{a \code{varfunc.rob} object}
\item{rho}{a parametric correlation function}
\item{method}{a method used for obtaining a mean and variance function}
}
\description{
Covariance estimation via semiparametric approach
}
