% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_Mest.R
\name{cv.cov_Mest}
\alias{cv.cov_Mest}
\title{K-fold Cross-Validation for Bivariate Nadaraya-Watson smoothing for Covariance}
\usage{
cv.cov_Mest(x, bw_cand = NULL, K = 5, ncores = 1, noise.var = 0)
}
\arguments{
\item{x}{a n x p matrix}

\item{bw_cand}{a numeric vector of bandwidth candidate. (Default is \code{NULL}, and it is selected automatically.)}

\item{K}{a number of folds for K-fold cross-validation. (Default is 5.)}

\item{ncores}{a number of cores to select bandwidth from robust 5-fold cross-validation. (Defalut is 1.)}

\item{noise.var}{the numeric value containing noise variance. (Defalut is 0.)}
}
\value{
a list contatining as follows:
\item{selected_bw}{the optimal bandwidth selected from the robust K-fold cross-validation.}
\item{cv.error}{a matrix containing CV error per bandwidth candidates.}
}
\description{
Perform K-fold cross-validation for Bivariate Nadaraya-Watson smoothing for Covariance.
}
