% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_ogk.R
\name{cv.cov_ogk}
\alias{cv.cov_ogk}
\title{K-fold cross-validation for Bivariate Nadaraya-Watson smoothing for Covariance}
\usage{
cv.cov_ogk(
  X,
  type = c("huber", "bisquare", "tdist"),
  bw_cand = NULL,
  K = 5,
  ncores = 1
)
}
\arguments{
\item{X}{a n x p matrix with or without NA.}

\item{type}{the option for robust dispersion estimator. "huber", "bisquare", and "tdist" are supported.}

\item{bw_cand}{a vector contains the candidates of bandwidths for bivariate smoothing.}

\item{K}{the number of folds for K-fold cross validation.}

\item{ncores}{the number of cores on \code{foreach} for parallel computing.}
}
\value{
a list contatining as follows:
\item{selected_bw}{the optimal bandwidth selected from the robust K-fold cross-validation.}
\item{cv.error}{a matrix containing CV error per bandwidth candidates.}
}
\description{
K-fold cross-validation for Bivariate Nadaraya-Watson smoothing for Covariance
}
\examples{
set.seed(100)
X <- sim_delaigle(n = 100,
                  type = "partial",
                  out.prop = 0.2,
                  out.type = 1,
                  dist = "normal")
X <- list2matrix(X)

# 5-fold CV for bivariate smoothing
bw_cand = seq(0.01, 0.1, length.out = 10)
cov.obj.cv <- cv.cov_ogk(X,
                         type = 'huber',
                         bw_cand = bw_cand,
                         K = 5)
print(cov.obj.cv$selected_bw)

# Robust smoothed covariance using selected bw from cross-validation
cov.obj <- cov_ogk(X,
                   type = "huber",
                   smooth = TRUE,
                   bw = cov.obj.cv$selected_bw)
mu.ogk.sm <- cov.obj$mean
cov.ogk.sm <- cov.obj$cov
noise.ogk.sm <- cov.obj$noise.var

}
\references{
\cite{Kim, H., Park, Y., & Lim, Y. (2022+). Robust principal component analysis and its application for partially observed functional data, Submitted.}
}
