% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dist.mat}
\alias{dist.mat}
\title{403 spatial locations and their distances matrix}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 403 rows and 403 columns.
}
\usage{
dist.mat
}
\description{
A dataset containing 403 spatial locations and their distances matrix
}
\keyword{datasets}
