% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{funPCA}
\alias{funPCA}
\title{Functional Principal Component Analysis (FPCA) via Conditional Expectation}
\usage{
funPCA(Lt, Ly, mu, cov, sig2 = 0, work.grid = NULL, K = NULL, PVE = 0.99)
}
\arguments{
\item{Lt}{a list of vectors or a vector containing time points for all curves}

\item{Ly}{a list of vectors or a vector containing observations for all curves}

\item{mu}{a mean function estimated at work.grid}

\item{cov}{a covariance function estimated at work.grid}

\item{sig2}{a noise variance}

\item{work.grid}{a work grid}

\item{K}{a number of FPCs. If K is NULL, K is selected by PVE.}

\item{PVE}{a proportion of variance explained}
}
\value{
a list contatining as follows:
\item{data}{a list containing Lt and Ly}
\item{lambda}{the first K eigenvalues}
\item{eig.fun}{the first K eigenvectors}
\item{pc.score}{the first K FPC scores}
\item{K}{a number of FPCs}
\item{PVE}{a proportion of variance explained}
\item{work.grid}{a work grid}
\item{eig.obj}{an object of the eigenanalysis}
\item{mu}{a mean function}
\item{cov}{a covariance function}
\item{sig2}{a noise variance}
}
\description{
FPCA is performed via PACE (Principal Analysis via Conditional Expectation) proposed by Yao et al. (2005).
}
