% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_corr.R
\name{get_corr_eigen}
\alias{get_corr_eigen}
\title{Get True eigenfunction for simulation of the spatially correlated case.}
\usage{
get_corr_eigen(grid)
}
\arguments{
\item{grid}{a vector containing the observed timepoints.}
}
\value{
a p x 4 matrix containing the eigenfunctions where p is the length of grids and 4 is the number of true components.
}
\description{
True eigenfunctions are obtained for given grids from the spatially correlated case.
}
\examples{
gr <- seq(0, 1, length.out = 51)
eig.true <- get_corr_eigen(gr)
matplot(eig.true, type = "l")

}
