% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_delaigle.R
\name{get_delaigle_cov}
\alias{get_delaigle_cov}
\title{Get true covariance function from Delaigle et al.(2021).}
\usage{
get_delaigle_cov(grid, model = 2)
}
\arguments{
\item{grid}{a vector containing the observed timepoints.}

\item{model}{The model option for Delaigle et al.(2021).}
}
\value{
a covariance matrix with the dimension, length of grids.
}
\description{
True covariance function is obtained for given grids from the setting of Delaigle et al.(2021) with model 2.
}
\examples{
gr <- seq(0, 1, length.out = 51)
cov.true <- get_delaigle_cov(gr, model = 2)
library(GA)
persp3D(gr, gr, cov.true,
        theta = -70, phi = 30, expand = 1)

}
\references{
\cite{Delaigle, A., Hall, P., Huang, W., & Kneip, A. (2021). Estimating the covariance of fragmented and other related types of functional data. Journal of the American Statistical Association, 116(535), 1383-1401.}
}
