% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_delaigle.R
\name{get_delaigle_eigen}
\alias{get_delaigle_eigen}
\title{Get True eigenfunction for simulation in Delaigle et al.(2021).}
\usage{
get_delaigle_eigen(grid, model = 2)
}
\arguments{
\item{grid}{a vector containing the observed timepoints.}

\item{model}{The model option for Delaigle et al.(2021).}
}
\value{
a p x k matrix containing the eigenfunctions where p is the length of grids and k is the number of true components.
}
\description{
True eigenfunctions are obtained for given grids from the setting of Delaigle et al.(2021) with model 2.
}
\examples{
gr <- seq(0, 1, length.out = 51)
eig.true <- get_delaigle_eigen(gr, model = 2)
matplot(eig.true, type = "l")

}
\references{
\cite{Delaigle, A., Hall, P., Huang, W., & Kneip, A. (2021). Estimating the covariance of fragmented and other related types of functional data. Journal of the American Statistical Association, 116(535), 1383-1401.}
}
