% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utills.R
\name{get_deriv}
\alias{get_deriv}
\title{Calculate discrete derivatives}
\usage{
get_deriv(f_x, x)
}
\arguments{
\item{f_x}{a vector of f(x)}

\item{x}{a vector of x}
}
\value{
a vector containing derivatives of \code{f_x}
}
\description{
Calculate discrete derivatives
}
