% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_design_index}
\alias{get_design_index}
\title{Get design points}
\usage{
get_design_index(Lt)
}
\arguments{
\item{Lt}{a list of vectors containing time points for each curve}
}
\description{
Get design points
}
