% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen.R
\name{get_eigen}
\alias{get_eigen}
\title{Get eigen analysis results}
\usage{
get_eigen(cov, grid)
}
\arguments{
\item{cov}{a matrix of covariance matrix}

\item{grid}{a vector containing observed time points. (same with dimension of \code{cov})}
}
\description{
Get eigen analysis results
}
