% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utills.R
\name{get_ise}
\alias{get_ise}
\title{Calculate Intergrated Squared Errors (ISE)}
\usage{
get_ise(x, x_hat, grid)
}
\arguments{
\item{x}{vectors or matrices to compare (y-axis)}

\item{x_hat}{vectors or matrices to compare (y-axis)}

\item{grid}{corresponding observed grid (x-axis)}
}
\value{
a value of ISE
}
\description{
Calculate Intergrated Squared Errors (ISE)
}
