% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_kraus.R
\name{get_kraus_eigen}
\alias{get_kraus_eigen}
\title{Get True eigenfunction for simulation in Kraus(2015).}
\usage{
get_kraus_eigen(grid)
}
\arguments{
\item{grid}{a vector containing the observed timepoints.}
}
\value{
a p x k matrix containing the eigenfunctions where p is the length of grids and k is the number of true components.
}
\description{
True eigenfunctions are obtained for given grids from the setting of Kraus(2015).
}
\examples{
gr <- seq(0, 1, length.out = 51)
eig.true <- get_kraus_eigen(gr)
matplot(eig.true, type = "l")

}
\references{
\cite{Kraus, D. (2015). Components and completion of partially observed functional data. Journal of the Royal Statistical Society: Series B: Statistical Methodology, 777-801.}
}
