% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_local_M.R
\name{get_raw_cov}
\alias{get_raw_cov}
\title{Get Raw Covariances for Functional Data}
\usage{
get_raw_cov(X, mu = NULL, gr = NULL, diag = FALSE, engine = "C++")
}
\arguments{
\item{X}{a n x p matrix}

\item{mu}{a vector containing the mean function. (Default is \code{NULL}, and it is calculated automatically.)}

\item{gr}{a vector containing observed timepoints. (Default is \code{NULL}, and it is set to equal grid between 0 and 1.)}

\item{diag}{a logical value whether containing diagonal parts of the covariance. (Default is \code{FALSE}.)}

\item{engine}{the option for implementation. (Defalut is \code{C++}, and it is much faster than \code{R}.)}
}
\value{
a dataframe contatining as follows:
\item{cov}{the raw covariance $C(s,t)$ between time \code{s} and \code{t} for \code{ind}th curve}
\item{s}{time 1}
\item{t}{time 2}
\item{ind}{the index of each curve}
}
\description{
Obtain the raw covariances $C(s,t)$.
}
\details{
It provides 2 options for implementing R or C++.
}
