% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utills.R
\name{is.convex}
\alias{is.convex}
\title{Check whether vector is convex}
\usage{
is.convex(f_x, x)
}
\arguments{
\item{f_x}{a vector of f(x)}

\item{x}{a vector of x}
}
\value{
a logical vector indicating \code{f_x} is convex on each \code{x}
}
\description{
Check whether vector is convex
}
