% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utills.R
\name{list2matrix}
\alias{list2matrix}
\title{Convert a list to a matrix}
\usage{
list2matrix(X)
}
\arguments{
\item{X}{a list containing 2 list (Lt, Ly).}
}
\value{
a n x p matrix containing functional trajectories.
n is the number of curves, and p is the number of timepoints.
}
\description{
Convert a list to a matrix
}
