% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utills.R
\name{list2rbind}
\alias{list2rbind}
\title{Rbind the list containing matrix having same number of columns}
\usage{
list2rbind(x)
}
\arguments{
\item{x}{a list containing matrices having same number of columns}
}
\value{
a matrix concatenating all matrices row-wise in a list
}
\description{
Rbind the list containing matrix having same number of columns
}
