% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRLScpp.R
\name{local_kern_smooth}
\alias{local_kern_smooth}
\title{Local polynomial kernel smoothing with huber loss (mean estimator)}
\usage{
local_kern_smooth(
  Lt,
  Ly,
  newt = NULL,
  method = c("L2", "HUBER", "WRM", "BISQUARE"),
  kernel = "epanechnikov",
  bw = NULL,
  delta = 1.345,
  deg = 1,
  ncores = 1,
  ...
)
}
\arguments{
\item{Lt}{a list of vectors or a vector containing time points for all curves}

\item{Ly}{a list of vectors or a vector containing observations for all curves}

\item{newt}{a vector containing time points to estimate}

\item{method}{"Huber" or "WRM" or "Bisquare"}

\item{kernel}{a kernel function for kernel smoothing ("epan", "gauss" are supported.)}

\item{bw}{bandwidth}

\item{delta}{cut-off value for Huber function}

\item{deg}{a degree of polynomial}

\item{ncores}{number of cores for k-fold cross-validation
For loss = "Huber", it uses \code{rlm()} in \code{MASS} and fits the robust regression with Huber loss.
So additional parameters of \code{rlm()} can be applied. (delta, maxit, ...)}

\item{...}{additional parameters}
}
\value{
a vector containing the smoothed functional trajectory from the local polynomial smoothing
}
\description{
Robust local polynomial smoothing via Huber function using iteratively re-weighted least sqruares (IRLS) algorithm
}
