% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRLScpp.R
\name{locpolysmooth}
\alias{locpolysmooth}
\title{Local polynomial kernel smoothing with huber loss (mean estimator)}
\usage{
locpolysmooth(
  Lt,
  Ly,
  newt = NULL,
  method = c("L2", "HUBER", "WRM", "BISQUARE"),
  kernel = "epanechnikov",
  bw = NULL,
  delta = NULL,
  deg = 1,
  cv = FALSE,
  ncores = 1,
  ...
)
}
\arguments{
\item{Lt}{a list of vectors or a vector containing time points for all curves}

\item{Ly}{a list of vectors or a vector containing observations for all curves}

\item{newt}{a vector containing time points to estimate}

\item{method}{"Huber" or "WRM" or "Bisquare"}

\item{kernel}{a kernel function for kernel smoothing ("epan", "gauss" are supported.)}

\item{bw}{bandwidth, default is (max(Lt)-min(Lt))/5.}

\item{delta}{cut-off value for "huber"(Huber) or "bisquare"(Tukey's biweight function).
Default is 1.345 for "huber" and 4.685 for "bisquare" for 95\% ARE.}

\item{deg}{a degree of polynomial}

\item{cv}{If TRUE, k-fold cross-validation is performed for bandwidth.}

\item{ncores}{number of cores for k-fold cross-validation}

\item{...}{additional parameters}
}
\value{
a vector containing the smoothed functional trajectory from the local polynomial smoothing
}
\description{
Robust local polynomial smoothing via Huber function using iteratively re-weighted least sqruares (IRLS) algorithm
}
