% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_local_M.R
\name{mean_local_M}
\alias{mean_local_M}
\title{Local M-estimator for Mean function (mean estimator)}
\usage{
mean_local_M(X, bw = NULL, gr = NULL, cv = FALSE, ncores = 1)
}
\arguments{
\item{X}{a n x p matrix}

\item{bw}{a numeric value of bandwidth. (Defalut is (max(gr) - min(gr)) / 5.)}

\item{gr}{a vector containing observed timepoints. (Default is \code{NULL}, and it is set to equal grid between 0 and 1.)}

\item{cv}{If it is \code{TRUE}, robust 5-fold cross-validation(using Huber loss) is performed to select the optimal bandwidth. (Default is \code{FALSE}.)}

\item{ncores}{a number of cores to select bandwidth from robust 5-fold cross-validation. (Defalut is 1.)}
}
\value{
a vector containing the mean function from local M-estimation.
}
\description{
Obtain the local M-estimator for mean function.
}
