% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_Mest.R
\name{mean_Mest}
\alias{mean_Mest}
\title{Marignal M-estimator for Mean function (mean estimator)}
\usage{
mean_Mest(x, smooth = FALSE)
}
\arguments{
\item{x}{a n x p matrix or a list containing Lt and Ly}

\item{smooth}{If it is \code{TRUE}, smoothing spline is performed in \code{stats::smooth.spline()}. (Default is \code{FALSE}.)}
}
\value{
a p-dimensional mean function
}
\description{
Obtain the marginal M-estimator for mean function.
}
\details{
It is performed by C++ code for fast computation.
}
