% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robfpca.partial.R
\name{plot.robfpca.partial}
\alias{plot.robfpca.partial}
\title{Plot robfpca.partial object}
\usage{
\method{plot}{robfpca.partial}(x, ...)
}
\arguments{
\item{x}{a \code{robfpca.partial} object from \code{robfpca.partial()}}

\item{...}{Not used}
}
\description{
Plot robfpca.partial object
}
