% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utills.R
\name{plot_curve}
\alias{plot_curve}
\title{Plot functional trajectories}
\usage{
plot_curve(x = NULL, Lt = NULL, Ly = NULL, ...)
}
\arguments{
\item{x}{a list containing 2 list (Lt, Ly) or n x p matrix.}

\item{Lt}{a list containing obeserved time grids of each trajectory. (If x is NULL)}

\item{Ly}{a list containing obeserved values of each trajectory. (If x is NULL)}

\item{...}{Options of \code{plot()}.}
}
\description{
Plot functional trajectories
}
