% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{pred_missing_curve}
\alias{pred_missing_curve}
\title{Obtain reconstructed curve for missing parts and simple curve registration}
\usage{
pred_missing_curve(x, pred, grid = NULL, align = FALSE, conti = TRUE)
}
\arguments{
\item{x}{a vector containing a partially observed curve (NA for not observed grids)}

\item{pred}{a vector containing a reconstructed curve for whole grid points}

\item{grid}{a vector containing grid points (if NULL, it uses equally spaced grids between (0, 1))}

\item{align}{If TRUE, a simple curve registration are performed for missing parts.}

\item{conti}{include last observed value, thus looks continuous prediction.}
}
\description{
Obtain reconstructed curve for missing parts and simple curve registration
}
