% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{predict.meanfunc.rob}
\alias{predict.meanfunc.rob}
\title{Predict mean at new time points}
\usage{
\method{predict}{meanfunc.rob}(object, newt, ...)
}
\arguments{
\item{object}{an \code{meanfunc.rob} object from \code{meanfunc.rob()}}

\item{newt}{a vector containing timepoints to predict}

\item{...}{does not needed}
}
\value{
a vector containing a mean function corresponds to \code{newt}
}
\description{
Predict mean at new time points
}
