% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robfpca.R
\name{predict.robfpca.partial}
\alias{predict.robfpca.partial}
\title{Predict FPC scores, reconstruction and completion for a new data}
\usage{
\method{predict}{robfpca.partial}(
  object,
  type = c("score", "reconstr", "comp"),
  newdata = NULL,
  K = NULL,
  ...
)
}
\arguments{
\item{object}{a \code{robfpca} object from \code{robfpca.partial()}}

\item{type}{"score" gives FPC scores, "reconstr" gives reconstruction of each curves, and "comp" gives completion of each curves.}

\item{newdata}{a n x p matrix containing n curves observed at p timepoints}

\item{K}{a number of FPCs}

\item{...}{Not used}
}
\description{
Predict FPC scores, reconstruction and completion for a new data
}
