% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var.R
\name{predict.varfunc.rob}
\alias{predict.varfunc.rob}
\title{Predict variance at new time points}
\usage{
\method{predict}{varfunc.rob}(object, newt, ...)
}
\arguments{
\item{object}{an object from \code{varfunc.rob()}}

\item{newt}{new time points to predict}

\item{...}{not used}
}
\description{
Predict variance at new time points
}
