% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robfpca.partial.R
\name{print.robfpca.partial}
\alias{print.robfpca.partial}
\title{Print a robfpca.partial object}
\usage{
\method{print}{robfpca.partial}(x, ...)
}
\arguments{
\item{x}{a \code{robfpca.partial} object from \code{robfpca.partial()}}

\item{...}{Not used}
}
\description{
Print a robfpca.partial object
}
