% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noise_var.R
\name{sigma2.rob}
\alias{sigma2.rob}
\title{Noise variance estimator}
\usage{
sigma2.rob(t, y, h = NULL)
}
\arguments{
\item{t}{a list of vectors containing time points for each curve}

\item{y}{a list of vectors containing observations for each curve}

\item{h}{a bandwidth}
}
\description{
Noise variance estimator
}
