% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_delaigle.R
\name{sim_delaigle}
\alias{sim_delaigle}
\title{Generate partially observed functional data from Delaigle et al.(2021).}
\usage{
sim_delaigle(
  n = 100,
  type = c("partial", "snippet", "dense"),
  out.prop = 0.2,
  out.type = 1,
  dist = "normal",
  noise = 0
)
}
\arguments{
\item{n}{a number of curves}

\item{type}{the type of generated data. "partial" means the option for partially observed data, "snippet" is short fragmented data, and "dense" means the fully observed curves.}

\item{out.prop}{a proportion of outlying curves of total n curves. Only used for dist = "normal".}

\item{out.type}{a outlier type, 1~3 are supported. Only used for dist = "normal".}

\item{dist}{a distribution which the data is generated. "normal"(Normal distribution) and "tdist"(t-distribution) are supported. If dist = "tdist", the option of \code{out.prop} and \code{out.type} are ignored.}

\item{noise}{a numeric value which is added random gaussian noises. Default is 0(No random noise).}
}
\value{
a n x 51 matrix with n observations per 51 timepoints
}
\description{
Partially observed functional data is generated with 51 regular grids by using the setting of Delaigle et al.(2021) with model 2.
}
\examples{
set.seed(100)
X <- sim_delaigle(n = 100,
                  type = "partial",
                  out.prop = 0.2,
                  out.type = 1,
                  dist = "normal")
X <- list2matrix(X)
matplot(t(X), type = "l")

}
\references{
\cite{Delaigle, A., Hall, P., Huang, W., & Kneip, A. (2021). Estimating the covariance of fragmented and other related types of functional data. Journal of the American Statistical Association, 116(535), 1383-1401.}
}
