% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_kraus.R
\name{sim_kraus}
\alias{sim_kraus}
\title{Generate partially observed functional data from Kraus. (2015).}
\usage{
sim_kraus(
  n = 100,
  type = c("partial", "snippet", "dense"),
  num.comp = 100,
  out.prop = 0.2,
  out.type = 1,
  dist = "normal",
  noise = 0
)
}
\arguments{
\item{n}{a number of curves}

\item{type}{the type of generated data. "partial" means the option for partially observed data, "snippet" is short fragmented data, and "dense" means the fully observed curves.}

\item{num.comp}{the number of components when the data are generated. See Kraus(2015).}

\item{out.prop}{a proportion of outlying curves of total n curves. Only used for dist = "normal".}

\item{out.type}{a outlier type, 1~3 are supported. Only used for dist = "normal".}

\item{dist}{a distribution which the data is generated. "normal"(Normal distribution) and "tdist"(t-distribution) are supported. If dist = "tdist", the option of \code{out.prop} and \code{out.type} are ignored.}

\item{noise}{a numeric value which is added random gaussian noises. Default is 0(No random noise).}
}
\value{
a list contatining as follows:
\item{Ly}{a list of n vectors containing the observed values for each individual.}
\item{Lt}{a list of n vectors containing the observation time points for each individual corresponding to \code{Ly}}
\item{out.ind}{a vector containing outlier index. 0 is non-outlier and 1 is the outlier.}
\item{x.full}{a n x 51 dense matrix with n observations per 51 timepoints before making partially observed.}
}
\description{
Partially observed functional data is generated with 51 regular grids by using the setting of Kraus(2015).
}
\examples{
set.seed(100)
x.list <- sim_kraus(n = 100,
                    type = "partial",
                    num.comp = 5,
                    out.prop = 0.2,
                    out.type = 1,
                    dist = "normal")
x <- list2matrix(x.list)
matplot(t(x), type = "l")

}
\references{
\cite{Kraus, D. (2015). Components and completion of partially observed functional data. Journal of the Royal Statistical Society: Series B: Statistical Methodology, 777-801.}
}
