% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var.R
\name{varfunc.rob}
\alias{varfunc.rob}
\title{Local polynomial kernel smoothing with huber loss (variance estimator)}
\usage{
varfunc.rob(
  Lt,
  Ly,
  newt = NULL,
  method = c("L2", "Huber", "WRM", "Bisquare"),
  mu = NULL,
  sig2 = NULL,
  ...
)
}
\arguments{
\item{Lt}{a list of vectors containing time points for each curve}

\item{Ly}{a list of vectors containing observations for each curve}

\item{newt}{a vector containing time points to estimate}

\item{method}{"huber", "WRM" are supported}

\item{mu}{\code{meanfunc.rob} object from \code{meanfunc.rob()}}

\item{sig2}{a noise variance estimator}

\item{...}{additional options of \code{meanfunc.rob()}}
}
\value{
a \code{varfunc.rob} object contatining \code{sig2} which is a noise variance estimate and \code{obj} which is containing as follows:
\item{t}{a vector containing unlist(Lt)}
\item{y}{a vector containing unlist(Ly)}
\item{n}{a number of functional trajectories}
\item{method}{a method used for obtaining a variance function}
\item{kernel}{a kernel used for local linear smoothing}
\item{bw}{a bandwidth}
\item{delta}{a cuf-off value in M-type loss function. ("HUBER")}
\item{deg}{a degree of polnomials}
\item{domain}{a range of timepoints}
\item{yend}{yend}
\item{cv_optns}{a list containing options of K-fold cross-validation. (\code{K} = \code{cv_K}, \code{ncore} = \code{ncores}, \code{delta_loss} = \code{cv_delta_loss}, \code{bw_loss} = \code{cv_bw_loss})}
}
\description{
Local polynomial kernel smoothing with huber loss (variance estimator)
}
