% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WRMcpp.R
\name{wrm_smooth}
\alias{wrm_smooth}
\title{Weighted repeated median (WRM) smoothing}
\usage{
wrm_smooth(x, y, h, xgrid = NULL, kernel = "epanechnikov")
}
\arguments{
\item{x}{a list of vectors containing time points for each curve}

\item{y}{a list of vectors containing observations for each curve}

\item{h}{bandwidth}

\item{xgrid}{a vector containing time points to estimate}

\item{kernel}{"epanechnikov" and "gauss" are supported.}
}
\value{
a vector containing the smoothed functional trajectory from weighted repeated median (WRM) smoothing
}
\description{
It is the same algorithm of \code{wrm.smooth()} in \code{robfilter} pacakges.
For fast computation, the function is written in C++.
The parameters of a function are the same with \code{wrm.smooth()} except on \code{kernel}.
}
\references{
\cite{Fried, R., Einbeck, J., & Gather, U. (2007). Weighted repeated median smoothing and filtering. Journal of the American Statistical Association, 102(480), 1300-1308.}
}
