% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_ogk.R
\name{cov_ogk}
\alias{cov_ogk}
\title{The robust covariance esimation based on modified Orthogonalized Gnanadesikan-Kettenring (OGK) estimation for partially observed functional data}
\usage{
cov_ogk(
  X,
  type = c("huber", "bisquare", "tdist"),
  smooth = FALSE,
  bw = NULL,
  noise.var = FALSE,
  df = 3,
  reweight = FALSE,
  beta = 0.9
)
}
\arguments{
\item{X}{a n x p matrix with or without NA.}

\item{type}{the option for robust dispersion estimator. "huber", "bisquare", and "tdist" are supported.}

\item{smooth}{If it is TRUE, bivariate Nadaraya-Watson smoothing is performed using \code{fields::smooth2d()}. Default is FALSE.}

\item{bw}{a bandwidth when \code{smooth = TRUE}.}

\item{noise.var}{If it is TRUE, we adjust the noise variance by using Yao et al.(2005). Default is FALSE.}

\item{df}{the degrees of freedm when \code{type = "tdist"}.}

\item{reweight}{If it is TRUE, the reweight procedure is performed in Maronna and Zammar(2002). Default is FALSE.}

\item{beta}{The parameter for reweight step. See Maronna and Zammar(2002).}
}
\value{
a list contatining as follows:
\item{mean}{the vector containing the robust mean function.}
\item{cov}{a matrix containing robust covariance function.}
\item{noise.var}{a noise variance estimator.}
}
\description{
The mean function is obtained via cross-sectional M-estimator of location.
The covariance function is obtained via OGK estimation using M-estimator of dispersion.
}
\examples{
set.seed(100)
X <- sim_delaigle(n = 100,
                  type = "partial",
                  out.prop = 0.2,
                  out.type = 1,
                  dist = "normal")
X <- list2matrix(X)
cov.obj <- cov_ogk(X,
                   type = "huber",
                   smooth = TRUE,
                   bw = 0.1)
mu.ogk.sm <- cov.obj$mean
cov.ogk.sm <- cov.obj$cov
noise.ogk.sm <- cov.obj$noise.var

}
\references{
\cite{Kim, H., Park, Y., & Lim, Y. (2022+). Robust principal component analysis and its application for partially observed functional data, Submitted.}

\cite{Maronna, R. A., & Zamar, R. H. (2002). Robust estimates of location and dispersion for high-dimensional datasets. Technometrics, 44(4), 307-317.}

\cite{Yao, F., Müller, H. G., & Wang, J. L. (2005). Functional data analysis for sparse longitudinal data. Journal of the American statistical association, 100(470), 577-590.}
}
