% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov.R
\name{covfunc.rob}
\alias{covfunc.rob}
\title{Covariance estimation via semiparametric approach}
\usage{
covfunc.rob(
  Lt,
  Ly,
  newt = NULL,
  method = c("Huber", "WRM", "Bisquare"),
  mu = NULL,
  sig2x = NULL,
  sig2e = NULL,
  corf = NULL,
  kernel = "epanechnikov",
  bw = NULL,
  delta = 1.345,
  deg = 1,
  cv = FALSE,
  ncores = 1,
  cv_bw_loss = "HUBER",
  cv_K = 5,
  ...
)
}
\arguments{
\item{Lt}{a list of vectors containing time points for each curve}

\item{Ly}{a list of vectors containing observations for each curve}

\item{newt}{a vector containing time points to estimate}

\item{method}{"huber", "WRM" are supported}

\item{mu}{Optional, \code{meanfunc.rob} object from \code{meanfunc.rob()}}

\item{sig2x}{Optional, \code{varfunc.rob} object from \code{varfunc.rob()}}

\item{sig2e}{Optional, a noise variance estimator obtained from \code{sigma2.rob()}}

\item{corf}{a correlation structure, it should be "function" class(x, y). Default is matern correlation.}

\item{kernel}{a kernel function for local polynomial smoothing ("epanechnikov", "gauss" are supported.)}

\item{bw}{a bandwidth for local polynomial smoothing.}

\item{delta}{If method == "Huber", it uses for $\\rho$ function in Huber loss.}

\item{deg}{a numeric scalar of polynomial degrees for local polynomial smoothing}

\item{cv}{If TRUE, K-fold cross-validation is performed for bandwidth.}

\item{ncores}{a number of cores to implement \code{foreach()} in \code{doParallel} for K-fold cross-validation.}

\item{cv_bw_loss}{a loss function for K-fold cross-validation}

\item{cv_K}{a number of folds for K-fold cross-validation}

\item{...}{additional options}
}
\value{
a \code{covfunc.rob} object contatining as follows:
\item{sig2}{a noise variance estimate}
\item{theta}{a estiamted parameter in parametric correlation function}
\item{mu.hat}{a estimated mean function corresponded to unlist(Lt)}
\item{domain}{a range of timepoints}
\item{mu}{a \code{meanfunc.rob} object}
\item{sig2x}{a \code{varfunc.rob} object}
\item{rho}{a parametric correlation function}
\item{method}{a method used for obtaining a mean and variance function}
}
\description{
Covariance estimation via semiparametric approach
}
