% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov.R
\name{covfunc.rob}
\alias{covfunc.rob}
\title{Covariance estimation via semiparametric approach}
\usage{
covfunc.rob(
  Lt,
  Ly,
  newt = NULL,
  method = c("Huber", "WRM", "Bisquare"),
  mu = NULL,
  weig = NULL,
  ...
)
}
\arguments{
\item{Lt}{a list of vectors containing time points for each curve}

\item{Ly}{a list of vectors containing observations for each curve}

\item{newt}{a vector containing time points to estimate}

\item{method}{"huber", "WRM" are supported}

\item{mu}{\code{meanfunc.rob} object from \code{meanfunc.rob()}}

\item{weig}{weight}

\item{...}{additional options of \code{varfunc.rob()}}
}
\description{
Covariance estimation via semiparametric approach
}
