% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{funPCA}
\alias{funPCA}
\title{PCA for functional snippets via conditional expectation}
\usage{
funPCA(Lt, Ly, mu, cov, sig2 = NULL, work.grid, K = NULL, PVE = 0.99)
}
\arguments{
\item{Lt}{a list of vectors or a vector containing time points for all curves}

\item{Ly}{a list of vectors or a vector containing observations for all curves}

\item{mu}{a mean function estimated at work.grid}

\item{cov}{a covariance function estimated at work.grid}

\item{sig2}{a noise variance}

\item{work.grid}{a work grid}

\item{K}{a number of PCs. If K is NULL, K is selected by PVE.}

\item{PVE}{a proportion of variance explained}
}
\description{
PCA for functional snippets via conditional expectation
}
