% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utills.R
\name{get_design_index}
\alias{get_design_index}
\title{Get design points}
\usage{
get_design_index(Lt)
}
\arguments{
\item{Lt}{a list of vectors containing time points for each curve}
}
\value{
a matrix containing 2 columns (each row is a point which is a pair of observed timepoints)
}
\description{
Get design points
}
