% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{meanfunc.rob}
\alias{meanfunc.rob}
\title{Local polynomial kernel smoothing with huber loss (mean estimator)}
\usage{
meanfunc.rob(
  Lt,
  Ly,
  newt = NULL,
  method = c("L2", "Huber", "WRM", "Bisquare"),
  kernel = "epanechnikov",
  bw = NULL,
  delta = NULL,
  deg = 1,
  ncores = 1,
  cv_delta_loss = "L1",
  cv_bw_loss = "HUBER",
  cv_K = 5,
  ...
)
}
\arguments{
\item{Lt}{a list of vectors containing time points for each curve}

\item{Ly}{a list of vectors containing observations for each curve}

\item{newt}{a vector containing time points to estimate}

\item{method}{"huber", "WRM" are supported}

\item{kernel}{a kernel function for kernel smoothing ("epanechnikov", "gauss" are supported.)}

\item{bw}{a bandwidth.
If bw = NULL, it is selected from K-fold cross-validation.}

\item{delta}{If method == "Huber", it uses for $\\rho$ function in Huber loss.
If delta = NULL, it is selected from K-fold cross-validation.}

\item{deg}{a numeric scalar of polynomial degrees for the kernel smoother}

\item{ncores}{a number of cores to implement \code{foreach()} in \code{doParallel} for K-fold cross-validation.}

\item{cv_delta_loss}{a loss function for K-fold cross-validation for delta in Huber function}

\item{cv_bw_loss}{a loss function for K-fold cross-validation for bandwidth}

\item{cv_K}{a number of folds for K-fold cross-validation}

\item{...}{additional options}
}
\value{
a \code{meanfunc.rob} object contatining follows:
\item{t}{a vector containing unlist(Lt)}
\item{y}{a vector containing unlist(Ly)}
\item{n}{a number of functional trajectories}
\item{method}{a method used for obtaining a mean function}
\item{kernel}{a kernel used for local linear smoothing}
\item{bw}{a bandwidth}
\item{delta}{a cuf-off value in M-type loss function. ("HUBER")}
\item{deg}{a degree of polnomials}
\item{domain}{a range of timepoints}
\item{yend}{yend}
\item{cv_optns}{a list containing options of K-fold cross-validation. (\code{K} = \code{cv_K}, \code{ncore} = \code{ncores}, \code{delta_loss} = \code{cv_delta_loss}, \code{bw_loss} = \code{cv_bw_loss})}
}
\description{
Local polynomial kernel smoothing with huber loss (mean estimator)
}
