% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov.R
\name{predict.covfunc.rob}
\alias{predict.covfunc.rob}
\title{Predict covariance at new time points}
\usage{
\method{predict}{covfunc.rob}(object, newt, ...)
}
\arguments{
\item{object}{an object from \code{covfunc.rob()}}

\item{newt}{new time points to predict}

\item{...}{not used}
}
\description{
Predict covariance at new time points
}
