% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{predict.meanfunc.rob}
\alias{predict.meanfunc.rob}
\title{Predict mean at new time points}
\usage{
\method{predict}{meanfunc.rob}(object, newt, ...)
}
\arguments{
\item{object}{an object from \code{meanfunc.rob()}}

\item{newt}{new time points to predict}

\item{...}{not used}
}
\description{
Predict mean at new time points
}
