% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var.R
\name{predict.varfunc.rob}
\alias{predict.varfunc.rob}
\title{Predict variance at new time points}
\usage{
\method{predict}{varfunc.rob}(object, newt, ...)
}
\arguments{
\item{object}{an object from \code{varfunc.rob()}}

\item{newt}{a vector containing timepoints to predict}

\item{...}{does not needed}
}
\value{
a vector containing a variance function corresponds to \code{newt}
}
\description{
Predict variance at new time points
}
