% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robfpca-package.R
\docType{package}
\name{robfpca-package}
\alias{robfpca}
\alias{robfpca-package}
\title{robfpca: Functional principal component analysis for partially observed elliptical process}
\description{
Functional principal component analysis for partially observed elliptical process Use four spaces when indenting paragraphs within the Description.
}
\author{
\strong{Maintainer}: Hyunsung Kim \email{hyunsung1021@gmail.com}

}
\keyword{internal}
