% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robfpca.R
\name{robfpca.partial}
\alias{robfpca.partial}
\title{Robust functional principal component analysis for partially observed functional data}
\usage{
robfpca.partial(
  X,
  type = c("huber", "bisquare", "tdist"),
  grid = NULL,
  K = NULL,
  PVE = 0.99,
  MM = TRUE,
  smooth = TRUE,
  bw = NULL,
  cv = FALSE,
  noise.var = FALSE,
  df = 3,
  cv_optns = list(bw_cand = NULL, K = 5, ncores = 1)
)
}
\arguments{
\item{X}{a n x p matrix. It allows NA.}

\item{type}{the option for robust dispersion estimator. "huber", "bisquare", and "tdist" are supported.}

\item{grid}{a vector containing the observed timepoints}

\item{K}{a number of FPCs. If K is NULL, K is selected by PVE.}

\item{PVE}{a proportion of variance explained}

\item{MM}{the option for M-scale estimator in GK identity. If it is FALSE, the same method using \code{type} is used, that is the iterative algorithm. The closed form solution using method of moments can be used when \code{MM == TRUE}. Defalut is TRUE.}

\item{smooth}{If it is TRUE, bivariate Nadaraya-Watson smoothing is performed using \code{fields::smooth2d()}. Default is TRUE.}

\item{bw}{a bandwidth when \code{smooth = TRUE}.}

\item{cv}{If it is TRUE, K-fold cross-validation is performed for the bandwidth selection when \code{smooth = TRUE}.}

\item{noise.var}{If it is TRUE, we adjust the noise variance by using Yao et al.(2005). Default is FALSE.}

\item{df}{the degrees of freedm when \code{type = "tdist"}.}

\item{cv_optns}{the options of K-fold cross-validation when \code{cv = TRUE}. See Details.}
}
\value{
a list contatining as follows:
\item{data}{a matrix which is the input X}
\item{lambda}{the first K eigenvalues}
\item{eig.fun}{the first K eigenvectors}
\item{pc.score}{the first K FPC scores}
\item{K}{a number of FPCs}
\item{PVE}{a proportion of variance explained}
\item{work.grid}{a work grid}
\item{eig.obj}{an object of the eigenanalysis}
\item{mu}{a mean function}
\item{cov}{a covariance function}
\item{sig2}{a noise variance}
\item{cov.obj}{the object from cov_ogk(). See ?cov_ogk().}
}
\description{
Robust functional principal component analysis (FPCA) for partially observed functional data.
It is based on the pairwise robust covariance function estimation and eigenanalysis.
}
\details{
The location and scale functions are computed via pointwise M-estimator, and the covariance function is obtained via robust pairwise computation based on Orthogonalized Gnanadesikan-Kettenring (OGK) estimation.
Additionally, bivariate Nadaraya-Watson smoothing is applied for smoothed covariance surfaces.
To deal with the missing segments, FPCA is performed via PACE (Principal Analysis via Conditional Expectation).

The options of \code{cv_optns}:
\describe{
\item{bw_cand}{a vector contains the candidates of bandwidths for bivariate smoothing.}
\item{K}{the number of folds for K-fold cross validation.}
\item{ncores}{the number of cores on \code{foreach} for parallel computing.}
}
}
\examples{
set.seed(100)
x.list <- sim_delaigle(n = 100,
                       type = "partial",
                       out.prop = 0.2,
                       out.type = 1,
                       dist = "normal")
x <- list2matrix(x.list)

# Given bandwidth
fpca.obj <- robfpca.partial(x,
                            type = "huber",
                            PVE = 0.95,
                            bw = 0.1)
fpc.score <- fpca.obj$pc.score

# Using 5-fold cross-validation
bw_cand <- seq(0.02, 0.3, length.out = 10)
fpca.obj <- robfpca.partial(x,
                            type = "huber",
                            PVE = 0.95,
                            cv = TRUE,
                            cv_optns = list(bw_cand = bw_cand,
                                            K = 5,
                                            ncores = 1))
fpc.score <- fpca.obj$pc.score

}
\references{
\cite{Park, Y., Kim, H., & Lim, Y. (2022+). Functional principal component analysis for partially observed elliptical process, Under review.}

\cite{Maronna, R. A., & Zamar, R. H. (2002). Robust estimates of location and dispersion for high-dimensional datasets. Technometrics, 44(4), 307-317.}

\cite{Yao, F., Müller, H. G., & Wang, J. L. (2005). Functional data analysis for sparse longitudinal data. Journal of the American statistical association, 100(470), 577-590.}
}
