% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var.R
\name{varfunc.rob}
\alias{varfunc.rob}
\title{Local polynomial kernel smoothing with huber loss (variance estimator)}
\usage{
varfunc.rob(
  Lt,
  Ly,
  newt = NULL,
  method = c("L2", "huber", "bisquare"),
  mu = NULL,
  sig2 = NULL,
  kernel = "epanechnikov",
  bw = NULL,
  delta = NULL,
  deg = 1,
  cv = FALSE,
  ncores = 1,
  cv_bw_loss = NULL,
  cv_K = 5,
  ...
)
}
\arguments{
\item{Lt}{a list of vectors containing time points for each curve}

\item{Ly}{a list of vectors containing observations for each curve}

\item{newt}{a vector containing time points to estimate}

\item{method}{"huber", "WRM" are supported}

\item{mu}{\code{meanfunc.rob} object from \code{meanfunc.rob()}}

\item{sig2}{a noise variance estimator obtained from \code{sigma2.rob()}}

\item{kernel}{a kernel function for local polynomial smoothing ("epanechnikov", "gauss" are supported.)}

\item{bw}{a bandwidth.}

\item{delta}{cut-off value for "huber"(Huber) or "bisquare"(Tukey's biweight function).
Default is 1.345 for "huber" and 4.685 for "bisquare" for 95\% ARE.}

\item{deg}{a numeric scalar of polynomial degrees for local polynomial smoothing}

\item{cv}{If TRUE, K-fold cross-validation is performed.}

\item{ncores}{a number of cores to implement \code{foreach()} in \code{doParallel} for K-fold cross-validation.}

\item{cv_bw_loss}{a loss function for K-fold cross-validation}

\item{cv_K}{a number of folds for K-fold cross-validation}

\item{...}{additional options}
}
\value{
a \code{varfunc.rob} object contatining \code{sig2} which is a noise variance estimate and \code{obj} which is containing as follows:
\item{t}{a vector containing unlist(Lt)}
\item{y}{a vector containing unlist(Ly)}
\item{n}{a number of functional trajectories}
\item{method}{a method used for obtaining a variance function}
\item{kernel}{a kernel used for local linear smoothing}
\item{bw}{a bandwidth}
\item{delta}{a cuf-off value in M-type loss function. ("HUBER")}
\item{deg}{a degree of polnomials}
\item{domain}{a range of timepoints}
\item{yend}{yend}
\item{cv_obj}{an object of K-fold CV for bandwidth}
\item{cv_optns}{a list containing options of K-fold cross-validation. (\code{K} = \code{cv_K}, \code{ncore} = \code{ncores}, \code{delta_loss} = \code{cv_delta_loss}, \code{bw_loss} = \code{cv_bw_loss})}
}
\description{
Local polynomial kernel smoothing with huber loss (variance estimator)
}
