% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var.R
\name{varfunc.rob}
\alias{varfunc.rob}
\title{Local polynomial kernel smoothing with huber loss (variance estimator)}
\usage{
varfunc.rob(
  Lt,
  Ly,
  newt = NULL,
  method = c("L2", "Huber", "WRM", "Bisquare"),
  mu = NULL,
  sig2 = NULL,
  ...
)
}
\arguments{
\item{Lt}{a list of vectors containing time points for each curve}

\item{Ly}{a list of vectors containing observations for each curve}

\item{newt}{a vector containing time points to estimate}

\item{method}{"huber", "WRM" are supported}

\item{mu}{\code{meanfunc.rob} object from \code{meanfunc.rob()}}

\item{sig2}{a noise variance estimator}

\item{...}{additional options of \code{meanfunc.rob()}}
}
\description{
Local polynomial kernel smoothing with huber loss (variance estimator)
}
